% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{build_rsa_paths}
\alias{build_rsa_paths}
\title{Build Response Surface Analysis Paths}
\usage{
build_rsa_paths(
  data,
  perceiver_id,
  target_id,
  group_id,
  rating_x,
  rating_y,
  rating_z,
  design = NULL
)
}
\arguments{
\item{data}{The dataframe. It must contain columns for percever, target,
and group ID variables and X, Y, and Z rating variables. Note that X and Y
can be the same variable. It should be in long format such
that each row is a rating by a perceiver for a given target in a given group.}

\item{perceiver_id}{A quoted string with the name of
the column containing perceiver IDs.
Perceiver IDs should be recycled across groups
(i.e., each group should have perceiver 1 to i
where i is the number of participants per group).
It can either be a number of a character string.}

\item{target_id}{A quoted string with the name of
the column containing target IDs.
target IDs should be recycled across groups
(i.e., each group should have perceiver 1 to i
where i is the number of participants per group).
It can either be a number of a character string.}

\item{group_id}{A quoted string with the name of
the column containing group IDs.It can either
be a number of a character string.}

\item{rating_x}{A quoted string with the name of
the column that contains ratings for the
x variable in the RSA.}

\item{rating_y}{A quoted string with the name of
the column that contains ratings for the
y variable in the RSA. Note that this can be the same
variable as x as long as design is not pxp.}

\item{rating_z}{A quoted string with the name of
the column that contains ratings for the
z variable (the outcome/DV) in the RSA.}

\item{design}{A quoted string specifying the design
of the RSA. Valid entries include:
\describe{
\item{reciprocal}{X and Y are reciprocal
ratings for each dyad; this can be on the same variable (e.g.,
A(B) Liking & B(A) Liking) or on different variables (e.g.,
A(B) Liking & B(A) Meta-Liking)}
\item{pxp}{X and Y are two ratings from the same
perceiver rating the same target (e.g., A(B) Liking
and A(B) Meta-Liking). These have to be different variables.}
\item{pxps}{X is a perceiver's rating of a target
and y is the perceivers' self-report.
This can be on the same or different variables.}
\item{pxts}{X is a perceiver's rating of a target
and y is the targets' self-report.
This can be on the same or different variables.}
\item{psxts}{X is a perceiver's self-report
and y is the targets' self-report.
This can be on the same or different variables.}}}
}
\description{
This function generates a string object corresponding
to the response surface analysis paths. It requires
a dataframe that contains columns for percever, target,
and group ID variables and rating variables for the X, Y,
Z variables for a response surface analysis (Z ~ X * Y).
It does not usually need to be called directly,
but is used in other functions.
}
\keyword{analysis}
\keyword{model,}
\keyword{relations}
\keyword{response}
\keyword{social}
\keyword{surface}
