% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_rsa_paths}
\alias{check_rsa_paths}
\title{Check Response Surface Analysis Paths}
\usage{
check_rsa_paths(fitted_srm_rsa)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
printed RSA Paths (in lavaan syntax).}
}
\value{
A printed, concatenated, string that corresponds to the RSA Paths
that get combined with the SRM and passed to lavaan.
}
\description{
This function concatenates and prints the string RSA Paths. If you're familiar with
lavaan syntax, this can be used to make sure the RSA is specified correctly in the event
of unexpected results.
}
\keyword{RSA}
