% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_srm}
\alias{check_srm}
\title{Check Social Relations Model}
\usage{
check_srm(fitted_srm_rsa)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
printed SRM (in lavaan syntax).}
}
\value{
A printed, concatenated, string that corresponds to the SRM passed to lavaan.
}
\description{
This function concatenates and prints the string srm model. If you're familiar with
lavaan syntax, this can be used to make sure the SRM is specified correctly in the
event of unexpected results.
}
\keyword{model}
\keyword{relations}
\keyword{social}
