% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_srm_rsa}
\alias{check_srm_rsa}
\title{Check Social Relations Model Response Surface Analysis}
\usage{
check_srm_rsa(fitted_srm_rsa)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
printed RSA Paths (in lavaan syntax).}
}
\value{
A printed, concatenated, string that corresponds to the SRM RSA
that is passed to lavaan.
}
\description{
This function concatenates and prints the full SRM RSA model. If you're familiar with
lavaan syntax, this can be used to make sure the model is specified correctly in the event
of unexpected results.
}
\keyword{RSA}
