% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rsa_modelcomp_tbl}
\alias{rsa_modelcomp_tbl}
\title{RSA (vs. NULL) Model Comparison Table}
\usage{
rsa_modelcomp_tbl(
  fitted_srm_rsa,
  caption = "Comparing RSA to NULL (no Interaction or Polynomials) Model"
)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
model comparison table.}

\item{caption}{The caption for the table. Defaults to generic label.}
}
\value{
a wrapper for knitr::kable that extracts and tables
the model comparison part of the SRM RSA output.
 Contains AIC, BIC, Chi-squared, difference in
model Chi-Squared values, and the corresponding df and p value
for the comparison.
}
\description{
This function extracts just the model comparison table from a fitted SRM
RSA object.
}
\keyword{RSA}
