% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{srm_rsa_surface_plot}
\alias{srm_rsa_surface_plot}
\title{SRM RSA Surface Plot}
\usage{
srm_rsa_surface_plot(fitted_srm_rsa, ...)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
surface plot.}

\item{...}{Optional additional arguments passed directly to \link[RSA]{plotRSA}
For example, it can be used to set X, Y, or Z labels, axis-limits, etc.}
}
\value{
a wrapper for RSA::plotRSA() that generates 3-d surface plot based
on the SRMRSA model output.
}
\description{
This function extracts the beta coefficients and generates a
3-d surface plot using the plotRSA() function from the RSA library
}
\keyword{RSA}
