% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{srm_rsa_surface_plot}
\alias{srm_rsa_surface_plot}
\title{SRM RSA Surface Plot}
\usage{
srm_rsa_surface_plot(fitted_srm_rsa)
}
\arguments{
\item{fitted_srm_rsa}{The fitted SRM RSA object from which you want the
surface plot.}
}
\value{
a wrapper for knitr::kable that extracts and tables
the model comparison part of the SRM RSA output.
 Contains AIC, BIC, Chi-squared, difference in
model Chi-Squared values, and the corresponding df and p value
for the comparison.
}
\description{
This function extracts the beta coefficients and generates a
3-d surface plot using the plotRSA() function from the RSA library
}
\keyword{RSA}
